Fixes several warnings.
git-svn-id: https://ad-svn.informatik.uni-freiburg.de/forestmaps@168 6bfcf819-cacf-47a1-ae15-21b88a40d8fc

+++ cpp-module/src/DirectedGraph.cpp
@@ -48,5 +48,5 @@ void OffsetListGraph<SourceTargetCostArc>::from_string(string s) {
   std::sort(_arcList.begin(), _arcList.end(),
             CompareArcs<SourceTargetCostArc>());
   _offset = compute_offsets(_arcList, numNodes);
-  assert(numNodes == this->num_nodes());
+  assert(static_cast<size_t>(numNodes) == this->num_nodes());
 }
