+++ cpp-module/src/EdgeAttractivenessModel.cpp
@@ -69,7 +69,7 @@ void EdgeAttractivenessModel::distribute(
     for (auto it2 = shares.begin(); it2 != shares.end(); ++it2) {
       const int edgeIndex = it2->first;
       const float share = it2->second;
-      assert(edgeIndex < attractivenesses->size());
+      assert(static_cast<size_t>(edgeIndex) < attractivenesses->size());
       (*attractivenesses)[edgeIndex] += share * population;
     }
   }
@@ -116,7 +116,7 @@ vector<float> FloodingModel::compute_edge_attractiveness() {
     done++;
     if ((clock() - timestamp) / static_cast<float>(CLOCKS_PER_SEC) > 0.00001) {
       timestamp = clock();
-      printf("Progress: %d of %d, this is %5.1f%% \r\n",
+      printf("Progress: %zu of %zu, this is %5.1f%% \r\n",
              done, total, done * 100.f / total);
     }
   }
@@ -196,7 +196,7 @@ vector<float> ViaEdgeApproach::compute_edge_attractiveness() {
     done++;
     if ((clock() - timestamp) / CLOCKS_PER_SEC > 2) {
       timestamp = clock();
-      printf("Progress: %d of %d, this is %5.1f%% \r\n",
+      printf("Progress: %zu of %zu, this is %5.1f%% \r\n",
              done, total, done * 100.f / total);
     }
   }
